Changed openness of some countries
Changed openess in such a way that I consider the game could be played in a much better way. One gateway in West Africa (Nigeria), one gateway in East Africa (Kenya), Brazil as a huge country in South America without which transferring would just be weird, and some small islands in the pacific that make life much more easier, especially if you want to do some form of island hopping.