Build base without a link to it or to country
1. Removed the limit that you have to have a link to airport before building a base
2. Removed the limit of having to have a link to country to gain airline tittle if your countries are allied (level 4) or same country (level 5).
3. Removed the limit of building bases only to gateway airports when in status of Privileged airline if your countries are allied (level 4) or same country (level 5).

The first one makes sense because not all airports need to be connected to be a base. Example is if you have two very close airports, or two airports in the same country and don't need a link between them.
The second and third change were made to simulate a close relationship within common travel areas. Now, for example everywhere in the EU (or in other unions/alliances), the airline will have the same status and be able to build in any airport (not just  gateways). Since overseas territories are considered a seperate country from the main country in the game, but they still have home country relationship level, this allows them to truly function as the same country. Before this, you needed to work on the relationship with every single territory, which did not make sense and was frustrating, especially for smaller Pacific territories of US, France, etc. This is also amplified by the fact those are usually not gateway airports. This fixes all of that that.
All other requirements (like delegates) stay as they make sense.